Add uuid_base to all accessories
Currently, homebridge uses the accessory type and display name to generate the
UUID. That means that if two devices have the same display name in the house,
we will crash on startup. It also means that renaming a device will have all
of its HomeKit configuration (like which room it is in) lost.

This change will break many peoples' setups - devices will be put in the
default room, and they will need to re-organize them.