Refactor cover, fixing some bugs
Cover devices may or may not support setting their position (ie, an arbitrary
value between 0 and 100) instead of just opening or closing.

Before this change, it was assumed that rollershutters could have their
set, but garage doors could not. This is not correct.

This refactors the cover accessory to make this separation more clear. Instead
of if statements scattered through the code, we build a class hierarchy to
clearly separate garage doors and rollershutters - of the "binary" type, and
the type we previously assumed all were.

At the same time, I moved all of the logic for choosing the cover type into
the cover factory, instead of doing it all in index.js